created all Dto classes. Adjusted int attributes to double.

+++ mmss/src/main/java/ca/mcgill/ecse/mmss/model/Artefact.java
@@ -29,8 +29,8 @@ public class Artefact
   private String description;
   @Column(nullable = false)
   private boolean canLoan;
-  private int insuranceFee;
-  private int loanFee;
+  private double insuranceFee;
+  private double loanFee;
 
   //Artefact Associations
   @ManyToOne
@@ -42,7 +42,7 @@ public class Artefact
 
   public Artefact () {}
   
-  public Artefact(int aArtefactId, String aArtefactName, String aDescription, boolean aCanLoan, int aInsuranceFee, int aLoanFee)
+  public Artefact(int aArtefactId, String aArtefactName, String aDescription, boolean aCanLoan, double aInsuranceFee, double aLoanFee)
   {
     artefactId = aArtefactId;
     artefactName = aArtefactName;
@@ -88,7 +88,7 @@ public class Artefact
     return wasSet;
   }
 
-  public boolean setInsuranceFee(int aInsuranceFee)
+  public boolean setInsuranceFee(double aInsuranceFee)
   {
     boolean wasSet = false;
     insuranceFee = aInsuranceFee;
@@ -96,7 +96,7 @@ public class Artefact
     return wasSet;
   }
 
-  public boolean setLoanFee(int aLoanFee)
+  public boolean setLoanFee(double aLoanFee)
   {
     boolean wasSet = false;
     loanFee = aLoanFee;
@@ -104,7 +104,7 @@ public class Artefact
     return wasSet;
   }
 
-  public int getArtefactId()
+  public  int getArtefactId()
   {
     return artefactId;
   }
@@ -124,12 +124,12 @@ public class Artefact
     return canLoan;
   }
 
-  public int getInsuranceFee()
+  public double getInsuranceFee()
   {
     return insuranceFee;
   }
 
-  public int getLoanFee()
+  public double getLoanFee()
   {
     return loanFee;
   }
